NeighborlyAPI/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions==1.2.1
pymongo==3.10.1